refactor: 패키지 정리
- 이벤트와 관련된 패키지를 event 하위로 이동
