+++ src/main/java/com/somemore/global/redis/registrar/RedisListenerRegistrar.java
@@ -1,7 +1,7 @@
 package com.somemore.global.redis.registrar;
 
-import com.somemore.interestcenter.subscriber.RedisCreateRecruitBoardSubscriber;
-import com.somemore.notification.subscriber.RedisNotificationSubscriber;
+import com.somemore.interestcenter.event.subscriber.RedisCreateRecruitBoardSubscriber;
+import com.somemore.notification.event.subscriber.RedisNotificationSubscriber;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
