+++ src/main/java/com/somemore/interestcenter/event/converter/CreateRecruitBoardMessageConverter.java
@@ -1,4 +1,4 @@
-package com.somemore.interestcenter.converter;
+package com.somemore.interestcenter.event.converter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
