+++ src/main/java/com/somemore/interestcenter/event/handler/CreateRecruitBoardHandler.java
@@ -1,4 +1,4 @@
-package com.somemore.interestcenter.handler;
+package com.somemore.interestcenter.event.handler;
 
 import com.somemore.recruitboard.event.CreateRecruitBoardEvent;
