+++ src/main/java/com/somemore/interestcenter/event/subscriber/RedisCreateRecruitBoardSubscriber.java
@@ -1,7 +1,7 @@
-package com.somemore.interestcenter.subscriber;
+package com.somemore.interestcenter.event.subscriber;
 
-import com.somemore.interestcenter.converter.CreateRecruitBoardMessageConverter;
-import com.somemore.interestcenter.handler.CreateRecruitBoardHandler;
+import com.somemore.interestcenter.event.converter.CreateRecruitBoardMessageConverter;
+import com.somemore.interestcenter.event.handler.CreateRecruitBoardHandler;
 import com.somemore.recruitboard.event.CreateRecruitBoardEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.connection.Message;
