+++ src/main/java/com/somemore/notification/event/converter/NotificationMessageConverter.java
@@ -1,11 +1,11 @@
-package com.somemore.notification.converter;
+package com.somemore.notification.event.converter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.somemore.community.event.CommentAddedEvent;
 import com.somemore.facade.event.VolunteerReviewRequestEvent;
-import com.somemore.interestcenter.event.InterestCenterCreateRecruitBoardEvent;
+import com.somemore.interestcenter.event.domain.InterestCenterCreateRecruitBoardEvent;
 import com.somemore.notification.domain.Notification;
 import com.somemore.notification.domain.NotificationSubType;
 import com.somemore.volunteerapply.domain.ApplyStatus;
