+++ src/main/java/com/somemore/notification/event/handler/NotificationHandler.java
@@ -1,4 +1,4 @@
-package com.somemore.notification.handler;
+package com.somemore.notification.event.handler;
 
 import com.somemore.notification.domain.Notification;
