+++ src/test/java/com/somemore/interestcenter/handler/CreateRecruitBoardHandlerImplTest.java
@@ -2,7 +2,8 @@ package com.somemore.interestcenter.handler;
 
 import com.somemore.IntegrationTestSupport;
 import com.somemore.global.common.event.ServerEventPublisher;
-import com.somemore.interestcenter.event.InterestCenterCreateRecruitBoardEvent;
+import com.somemore.interestcenter.event.domain.InterestCenterCreateRecruitBoardEvent;
+import com.somemore.interestcenter.event.handler.CreateRecruitBoardHandlerImpl;
 import com.somemore.interestcenter.usecase.InterestCenterQueryUseCase;
 import com.somemore.recruitboard.event.CreateRecruitBoardEvent;
 import org.junit.jupiter.api.Test;
